update permission database

+++ routes/web.php
@@ -24,43 +24,41 @@ Route::middleware("auth")->group(function () {
 
     Route::get('/ncr', [NcrController::class, "index"]);
 
-    Route::get('/ncr/create', [NcrController::class, "create"]);
+    Route::get('/ncr/create', [NcrController::class, "create"])->middleware("permission:add-ncr");
 
-    Route::post('/ncr', [NcrController::class, "store"]);
+    Route::post('/ncr', [NcrController::class, "store"])->middleware("permission:add-ncr");
 
     Route::get('/ncr/{ncr}', [NcrController::class, "show"]);
 
-    Route::get('/ncr/{ncr}/edit', [NcrController::class, "edit"]);
+    Route::get('/ncr/{ncr}/edit', [NcrController::class, "edit"])->middleware("permission:edit-ncr");
 
-    Route::put('/ncr/{ncr}', [NcrController::class, "update"]);
+    Route::put('/ncr/{ncr}', [NcrController::class, "update"])->middleware("permission:edit-ncr");
 
-    Route::delete('/ncr/{ncr}', [NcrController::class, "destroy"]);
+    Route::delete('/ncr/{ncr}', [NcrController::class, "destroy"])->middleware("permission:delete-ncr");
 
     Route::get("/kontak", [KontakController::class, "index"]);
 
-    Route::get("/kontak/create", [KontakController::class, "create"]);
+    Route::get("/kontak/create", [KontakController::class, "create"])->middleware("permission:add-kontak");
 
-    Route::post("/kontak", [KontakController::class, "store"]);
+    Route::post("/kontak", [KontakController::class, "store"])->middleware("permission:add-kontak");
 
-    Route::get("/kontak/{kontak}", [KontakController::class, "show"]);
+    Route::get("/kontak/{kontak}/edit", [KontakController::class, "edit"])->middleware("permission:edit-kontak");
 
-    Route::get("/kontak/{kontak}/edit", [KontakController::class, "edit"]);
+    Route::put("/kontak/{kontak}", [KontakController::class, "update"])->middleware("permission:edit-kontak");
 
-    Route::put("/kontak/{kontak}", [KontakController::class, "update"]);
-
-    Route::delete("/kontak/{kontak}", [KontakController::class, "destroy"]);
+    Route::delete("/kontak/{kontak}", [KontakController::class, "destroy"])->middleware("permission:delete-kontak");
     
     Route::get("/memo", [MemoController::class, "index"]);
     
-    Route::get("/memo/{ncr}/create", [MemoController::class, "create"]);
+    Route::get("/memo/{ncr}/create", [MemoController::class, "create"])->middleware("permission:add-memo");
 
-    Route::get("/memo/{ncr}/edit", [MemoController::class, "edit"]);
+    Route::get("/memo/{ncr}/edit", [MemoController::class, "edit"])->middleware("permission:edit-memo");
 
-    Route::put("/memo/{ncr}", [MemoController::class, "update"]);
+    Route::put("/memo/{ncr}", [MemoController::class, "update"])->middleware("permission:edit-memo");
 
-    Route::delete("/memo/{ncr}", [MemoController::class, "destroy"]);
+    Route::delete("/memo/{ncr}", [MemoController::class, "destroy"])->middleware("permission:delete-memo");
 
-    Route::post("/memo/{ncr}", [MemoController::class, "store"]);
+    Route::post("/memo/{ncr}", [MemoController::class, "store"])->middleware("permission:add-memo");
 
     Route::post("/ncr/validasi/{ncr}", [NcrController::class, "validasi"]);
 
@@ -73,25 +71,6 @@ Route::middleware("auth")->group(function () {
     Route::get("/role", [RoleController::class, "index"])->middleware("role:Admin");
 
     Route::post("/role/{user}", [RoleController::class, "update"])->middleware("role:Admin");
-});
-
-Route::group(['middleware' => ['role:admin','permission:add-kontak|edit-kontak|delete-kontak|add-ncr|edit-ncr|delete-ncr|create-memo|edit-memo|delete-memo']], function () {
-    
-    Route::get("/memo/{ncr}/create", [MemoController::class, "create"]);
-
-    Route::get("/memo/{ncr}/edit", [MemoController::class, "edit"]);
-
-    Route::put("/memo/{ncr}", [MemoController::class, "update"]);
-
-    Route::delete("/memo/{ncr}", [MemoController::class, "destroy"]);
-
-    Route::post("/memo/{ncr}", [MemoController::class, "store"]);
-
-    Route::post("/ncr/validasi/{ncr}", [NcrController::class, "validasi"]);
-
-    Route::get("/role", [RoleController::class, "index"]);
-
-    Route::post("/role/{user}", [RoleController::class, "update"]);
 });
 
 Auth::routes(["register"=>false]);
